+++ New/TelegramWorkLibrary/TelegramWorkLibrary/Struct/Audio.cs
@@ -9,13 +9,13 @@ namespace TelegramWorkLibrary
     public class Audio
     {
         public String _fileId { get; set; } // ID файла
-        public int _duration { get; set; } // Продолжительность файл 
+        public Int64 _duration { get; set; } // Продолжительность файл 
 
         // Необязательные параметры 
         public String _performer { get; set; } // Исполнитель 
         public String _title { get; set; } // Заголовок файла 
         public String _mimeType { get; set; } // MIME файла
-        public int? _fileSize { get; set; } // Размер файла
+        public Int64? _fileSize { get; set; } // Размер файла
     }
 }
