+++ New/TelegramWorkLibrary/TelegramWorkLibrary/Struct/Sticker.cs
@@ -9,10 +9,10 @@ namespace TelegramWorkLibrary
     public class Sticker
     {
         public String _fileId { get; set; } // ID стикера 
-        public int _width { get; set; } // Ширина стикера 
-        public int _height { get; set; } // Высота стикера 
+        public Int64 _width { get; set; } // Ширина стикера 
+        public Int64 _height { get; set; } // Высота стикера 
 
         public PhotoSize _thumb { get; set; } // превью стикера 
-        public int? _fileSize { get; set; } // Размер файла
+        public Int64? _fileSize { get; set; } // Размер файла
     }
 }
