+++ New/TelegramWorkLibrary/TelegramWorkLibrary/Struct/Voice.cs
@@ -9,9 +9,9 @@ namespace TelegramWorkLibrary
     public class Voice
     {
         public String _fileId { get; set; } // Уникальный идентификатор файла 
-        public int _duration { get; set; } // Продолжительность аудиофайла, заданная отправителем
+        public Int64 _duration { get; set; } // Продолжительность аудиофайла, заданная отправителем
 
         public String _mimeType { get; set; } // Опционально. MIME-тип файла, заданный отправителем
-        public int? _fileSize { get; set; } // Опционально. Размер файла
+        public Int64? _fileSize { get; set; } // Опционально. Размер файла
     }
 }
